[CLEARINGHOUSE] #REFACTOR 'module:clearinghouse' {Structure refactoring to match Broker-Module structure.}
